Extracted special cases minimum mask rect for the cursor into a separate method.
Caching the minimum mask rect calculated directly from image data so it only has to be recalculated when the image changes.  It's not actually that expensive of an operation since cursor images are small, but there's no point in doing it again, if we're still using the same cursor image.